Move the data center picker to its own block
