+++ client/blocks/data-center-picker/README.md
@@ -1,0 +1,19 @@
+# Data Center Picker (TSX)
+
+This component displays a data center picker. It lets the user choose between our recommended data center or through a custom list of avaliable data centers.
+
+## How to use
+
+```js
+import DataCenterPicker from 'calypso/blocks/data-center-picker';
+
+function render() {
+	return <DataCenterPicker />;
+}
+```
+
+## Props
+
+- `onChange (func)` - A function to handle the event callback when choosing a data center or picking the recommended one.
+- `value (string)` - The picked data center (this is a controlled component).
+- `compact (boolean)` - Whether or not to display a version of the data center picker without the global map.
