+++ client/landing/stepper/declarative-flow/internals/steps-repository/site-options/index.tsx
@@ -9,7 +9,7 @@ import { useTranslate } from 'i18n-calypso';
 import React, { useEffect } from 'react';
 import siteOptionsUrl from 'calypso/assets/images/onboarding/site-options.svg';
 import storeImageUrl from 'calypso/assets/images/onboarding/store-onboarding.svg';
-import DataCenterPicker from 'calypso/blocks/eligibility-warnings/data-center-picker';
+import DataCenterPicker from 'calypso/blocks/data-center-picker';
 import DocumentHead from 'calypso/components/data/document-head';
 import FormattedHeader from 'calypso/components/formatted-header';
 import FormFieldset from 'calypso/components/forms/form-fieldset';
